로그인 및 메인 레이아웃에서 사용자 인증 로직 개선: 불필요한 상태 관리 코드 제거 및 리다이렉션 처리 추가, 새로운 스레드 저장소 삭제
