+++ app/main/quick/page.tsx
@@ -3,15 +3,5 @@ import { redirect } from "next/navigation";
 import { QuickThreadWrapper } from "./threads";
 
 export default async function QuickPage() {
-  const supabase = createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    redirect("/login");
-  }
-
   return <QuickThreadWrapper />;
 }
